core/agent_manager.py
import os
from typing import Dict, List, Optional
from dotenv import load_dotenv

from agents.legal_architect.agent import LegalArchitect
from agents.hr_strategist.agent import HRStrategist
from agents.financial_expert.agent import FinancialExpert
from agents.governance_officer.agent import GovernanceOfficer
from agents.pmo_director.agent import PMODirector
from agents.evaluation_analyst.agent import EvaluationAnalyst

class AgentManager:
    """Orchestrates the interaction between different AI agents in the LLC Generator system."""
    
    def __init__(self):
        load_dotenv()
        self.agents = self._initialize_agents()
        self.workflow_state: Dict = {}
    
    def _initialize_agents(self) -> Dict:
        """Initialize all specialized AI agents."""
        return {
            "legal": LegalArchitect(api_key=os.getenv("OPENAI_API_KEY")),
            "hr": HRStrategist(api_key=os.getenv("ANTHROPIC_API_KEY")),
            "financial": FinancialExpert(api_key=os.getenv("GROQ_API_KEY")),
            "governance": GovernanceOfficer(api_key=os.getenv("GOOGLE_API_KEY")),
            "pmo": PMODirector(api_key=os.getenv("COHERE_API_KEY")),
            "evaluation": EvaluationAnalyst(api_key=os.getenv("EMERGENCEAI_API_KEY"))
        }
    
    async def process_llc_formation(self, business_details: Dict) -> Dict:
        """
        Orchestrate the LLC formation process through all agents.
        
        Args:
            business_details: Dictionary containing business information
            
        Returns:
            Dict containing the complete LLC formation package
        """
        # Step 1: Legal Structure and Compliance
        legal_structure = await self.agents["legal"].process(business_details)
        self.workflow_state["legal"] = legal_structure
        
        # Step 2: HR Setup and EIN Assignment
        hr_setup = await self.agents["hr"].process({
            **business_details,
            "legal_structure": legal_structure
        })
        self.workflow_state["hr"] = hr_setup
        
        # Step 3: Financial Setup
        financial_setup = await self.agents["financial"].process({
            **business_details,
            "legal_structure": legal_structure,
            "hr_setup": hr_setup
        })
        self.workflow_state["financial"] = financial_setup
        
        # Step 4: Governance Framework
        governance_framework = await self.agents["governance"].process({
            **business_details,
            "legal_structure": legal_structure,
            "hr_setup": hr_setup,
            "financial_setup": financial_setup
        })
        self.workflow_state["governance"] = governance_framework
        
        # Step 5: Project Management Setup
        pmo_setup = await self.agents["pmo"].process({
            **business_details,
            **self.workflow_state
        })
        self.workflow_state["pmo"] = pmo_setup
        
        # Step 6: Final Evaluation and Validation
        evaluation_result = await self.agents["evaluation"].process(self.workflow_state)
        
        if not evaluation_result["is_valid"]:
            # Trigger refinement loop
            return await self._handle_validation_failure(evaluation_result)
        
        return self._prepare_final_package(evaluation_result)
    
    async def _handle_validation_failure(self, evaluation_result: Dict) -> Dict:
        """Handle cases where the LLC formation package fails validation."""
        # Implement refinement loop logic here
        pass
    
    def _prepare_final_package(self, evaluation_result: Dict) -> Dict:
        """Prepare the final LLC formation package for delivery."""
        return {
            "legal_documents": self.workflow_state["legal"]["documents"],
            "hr_package": self.workflow_state["hr"]["package"],
            "financial_setup": self.workflow_state["financial"]["setup"],
            "governance_framework": self.workflow_state["governance"]["framework"],
            "pmo_setup": self.workflow_state["pmo"]["setup"],
            "evaluation_report": evaluation_result["report"],
            "next_steps": self._generate_next_steps()
        }
    
    def _generate_next_steps(self) -> List[Dict]:
        """Generate next steps for the LLC implementation."""
        return [
            {
                "phase": "Implementation",
                "tasks": [
                    "File LLC documents with state",
                    "Set up business banking",
                    "Implement HR systems",
                    "Configure PMO workflows",
                    "Deploy governance framework"
                ]
            }
        ]
