requirements.txt
langchain>=0.1.0
openai>=1.3.0
anthropic>=0.5.0
groq>=0.3.0
google-api-python-client>=2.0.0
cohere>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.20.0
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
aiohttp>=3.8.0
async-timeout>=4.0.0
